requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.49.0
numpy>=1.24.0
pandas>=2.0.0

# Data processing
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0

# Model-specific dependencies
# For Chronos-2
chronos-forecasting>=2.0.0
accelerate>=0.25.0

# For TimesFM
einops>=0.7.0

# For TiRex
tirex>=0.0.2
xlstm>=1.0.0

# Evaluation metrics
statsmodels>=0.14.0

# Optional: JAX support for TimesFM (if needed)
# jax>=0.4.20
# flax>=0.7.0
